refactored geoserver persister
