+++ backend/persisters/__init__.py
@@ -1,0 +1,10 @@
+# ===============================================================================
+# Author:  Jake Ross
+# Copyright 2025 New Mexico Bureau of Geology & Mineral Resources
+# Licensed under the Apache License, Version 2.0 (the "License");
+# You may not use this file except in compliance with the License.
+# You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
+# ===============================================================================
+
+
+# ============= EOF =============================================
